requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
qdrant-client==1.7.0
fastembed==0.1.3
python-multipart==0.0.6
python-dotenv==1.0.1